requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-interface==0.26.0
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-colorfield==0.11.0
django-import-export==3.2.0
djangoql==0.17.1
djangorestframework==3.15.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==10.2.0
mpmath==1.3.0
mssql-django==1.4
mysqlclient==2.2.4
networkx==3.3
numba==0.59.1
numpy==1.26.4
odfpy==1.4.1
openai-whisper==20231117
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
ply==3.11
pyodbc==5.1.0
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2023.3
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tablib==3.6.1
text-unidecode==1.3
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.41.2
xlrd==2.0.1
xlwt==1.3.0
